Use MUnit in SwaggerModelsBuilderSuite
